+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes to be released in next version
 
 🙌 Improvements
  * MXKEventFormatter: Expose defaultRoomSummaryUpdater ivar as protected.
+ * MXKCallViewController: Add transfer button and implement actions.
 
 🐛 Bugfix
  *